Rename constants. Move breakpoint types to same file. Add component export. Inline type imports in JS files.
