+++ src/Breakpoint/breakpoints.d.ts
@@ -3,3 +3,7 @@
  * @type {Record<BreakpointSize, BreakpointValue>}
  */
 export const breakpoints: Record<BreakpointSize, BreakpointValue>;
+
+export type BreakpointSize = "sm" | "md" | "lg" | "xlg" | "max";
+
+export type BreakpointValue = 320 | 672 | 1056 | 1312 | 1584;
