+++ src/Breakpoint/index.d.ts
@@ -1,2 +1,3 @@
+export { default as Breakpoint } from "./Breakpoint.svelte";
 export { breakpointObserver } from "./breakpointObserver";
-export { breakpoints } from "./constants";
+export { breakpoints } from "./breakpoints";
