+++ src/Breakpoint/index.js
@@ -1,3 +1,3 @@
 export { default as Breakpoint } from "./Breakpoint.svelte";
 export { breakpointObserver } from "./breakpointObserver";
-export { breakpoints } from "./constants";
+export { breakpoints } from "./breakpoints";
