--- src/Breakpoint/types.d.ts
@@ -1,2 +1,0 @@
-declare type BreakpointSize = "sm" | "md" | "lg" | "xlg" | "max";
-declare type BreakpointValue = 320 | 672 | 1056 | 1312 | 1584;
